requirements.txt
django>=3.0.3,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-debug-toolbar>=2.2.0,<2.3.0
